Split Maestro.Contracts so Darc can be used on non win7-x64 oses (#2161)
* Split Maestro.Contracts

* resolve comments

* Remove win7-x64 identifier from DarcLib.Tests
