+++ arcade-services.sln
@@ -209,7 +209,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Authentica
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Authentication.Algorithms.Tests", "src\Shared\Microsoft.DotNet.Authentication.Algorithms.Tests\Microsoft.DotNet.Authentication.Algorithms.Tests.csproj", "{8A8A95E8-C834-4771-9FF8-82A65F5E2BE5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests", "src\Microsoft.DotNet.Darc\tests\Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests\Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests.csproj", "{8B2E07FA-ED55-40B4-8D3B-F73B4BD641F8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests", "src\Microsoft.DotNet.Darc\tests\Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests\Microsoft.DotNet.Darc.VirtualMonoRepo.E2E.Tests.csproj", "{8B2E07FA-ED55-40B4-8D3B-F73B4BD641F8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Maestro.MergePolicyEvaluation", "src\Maestro\Maestro.MergePolicyEvaluation\Maestro.MergePolicyEvaluation.csproj", "{D16DB73B-4EDF-4E61-9392-238CF3800DE2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -1157,6 +1159,18 @@ Global
 		{8B2E07FA-ED55-40B4-8D3B-F73B4BD641F8}.Release|x64.Build.0 = Release|Any CPU
 		{8B2E07FA-ED55-40B4-8D3B-F73B4BD641F8}.Release|x86.ActiveCfg = Release|Any CPU
 		{8B2E07FA-ED55-40B4-8D3B-F73B4BD641F8}.Release|x86.Build.0 = Release|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Debug|x64.Build.0 = Debug|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Debug|x86.Build.0 = Debug|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Release|x64.ActiveCfg = Release|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Release|x64.Build.0 = Release|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Release|x86.ActiveCfg = Release|Any CPU
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1227,6 +1241,7 @@ Global
 		{98119800-E540-4928-8354-5E21E1384169} = {FB572E03-6F42-4262-ABB6-9D481E3647D6}
 		{8A8A95E8-C834-4771-9FF8-82A65F5E2BE5} = {FB572E03-6F42-4262-ABB6-9D481E3647D6}
 		{8B2E07FA-ED55-40B4-8D3B-F73B4BD641F8} = {C0B90032-033F-4C3E-809F-BD3339E1E313}
+		{D16DB73B-4EDF-4E61-9392-238CF3800DE2} = {AE791E26-78E1-4936-BCF4-1BF5152CBBD6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {32B9C883-432E-4FC8-A1BF-090EB033DD5B}
