+++ src/Maestro/Maestro.MergePolicies/MergePolicy.cs
@@ -5,6 +5,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using Microsoft.DotNet.DarcLib;
 using Newtonsoft.Json.Linq;
