+++ src/Maestro/Maestro.MergePolicies/StandardMergePolicy.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
