+++ src/Maestro/Maestro.MergePolicyEvaluation/IMergePolicyInfo.cs
@@ -2,12 +2,11 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Threading.Tasks;
-
-namespace Maestro.Contracts;
+namespace Maestro.MergePolicyEvaluation;
 
 public interface IMergePolicyInfo
 {
     string Name { get; }
     string DisplayName { get; }
 }
+
