+++ src/Maestro/Maestro.MergePolicyEvaluation/MergePolicyEvaluationStatus.cs
@@ -1,0 +1,12 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Maestro.MergePolicyEvaluation;
+
+public enum MergePolicyEvaluationStatus
+{
+    Pending = 0,
+    Success,
+    Failure,
+}
