+++ src/Microsoft.DotNet.Darc/src/Darc/Constants.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using System;
 using System.Collections.Generic;
 using System.IO;
