+++ src/Microsoft.DotNet.Darc/src/DarcLib/GitHubClient.cs
@@ -19,8 +19,8 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.DotNet.Services.Utility;
-using Maestro.Contracts;
 using System.Collections.Immutable;
+using Maestro.MergePolicyEvaluation;
 
 namespace Microsoft.DotNet.DarcLib;
