+++ src/Microsoft.DotNet.Darc/src/DarcLib/IRemote.cs
@@ -2,11 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using System.Xml;
-using Maestro.Contracts;
+using Maestro.MergePolicyEvaluation;
 using Microsoft.DotNet.Maestro.Client.Models;
 using NuGet.Versioning;
 using Asset = Microsoft.DotNet.Maestro.Client.Models.Asset;
